Fixed bug 1866166 - correction for norms and double args on 64 bit machines
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@3063 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/LAPACK.hpp
@@ -210,11 +210,11 @@ extern "C" {
 
   float snrm2_(int *N, float *X, int *INCX);
   
-  float dnrm2_(int *N, double *X, int *INCX);
+  double dnrm2_(int *N, double *X, int *INCX);
 
   float scnrm2_(int *N, float *X, int *INCX);
   
-  float dznrm2_(int *N, double *X, int *INCX);
+  double dznrm2_(int *N, double *X, int *INCX);
 
   void dgecon_(char *norm, int *N, double *A, int *LDA, double *Anorm,
 	       double *rcond, double *work, int *iwork, int *info);
